backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
